+++ flatlaf-natives/flatlaf-natives-windows/src/main/cpp/ApiVersion.cpp
@@ -24,7 +24,7 @@
 
 // increase this version if changing API or functionality of native library
 // also update version in Java class com.formdev.flatlaf.ui.FlatNativeWindowsLibrary
-#define API_VERSION_WINDOWS		1002
+#define API_VERSION_WINDOWS		1003
 
 
 //---- JNI methods ------------------------------------------------------------